requirements.txt
-f /usr/share/pip-wheels
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-extra-views==0.12.0
idna==2.8
Pillow==6.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
